first step: creating c64 mod
