+++ src/lib.rs
@@ -59,8 +59,6 @@ pub mod c64;
 pub mod cbm_kernal;
 #[cfg(feature = "cia")]
 pub mod cia;
-#[cfg(feature = "c64")]
-pub mod cpu6510;
 #[cfg(feature = "cx16")]
 pub mod cx16;
 #[cfg(feature = "mega65")]
